requirements.txt
streamlit
huggingface-hub==0.26.1
python-dotenv==1.0.1